🧹 `Journal`: Migrate Rendering to EntryComponent
- https://github.com/zinc-collective/convene/issues/1566

This refactor allows us to do post-processing of the markdown at the
Component level, which has access to the rails URL builders; which
should allow us to link to the `Keyword` within the `Entry#body`...
